Motor Speed Analysis
This is the last version of the file we downloaded from the link. The
file is manipulated such that not a single pulse is given to the motor
but a continuous signal which consists of pulses of whose period can be
adjusted with the Ts variable. The file gives an output of the speed of
the motor turned into rpm. From the scope we see that the speed is
neither constant nor smooth. We observe this output from the motor
itself as well, it rotates and stops continually due to the pwm it is
fed.
There is this fact when the code is built: There is a single pulse
first, then this single pulse is repeated. Single pulse's pwm ratio is
based on samples. And there is a term "sample ratio".  Only when this
ratio is changed the frequency of the whole signal can be changed.
However, the sample ratio of the encoder must be different (large in the
scales of 1 etc.) so that at the output scope, we see a smoother curve.
Here, I mean that normally a pulse whose frequency is quite high is
given but the encoder does not read at each changing step so that we see
an estimation (otherwise speed reading changes very much, does not look
well to the eye).
Overall, the speed analysis is satisfactory at the moment if you find
the motor's operation allright.